python-day6/requirements.txt
flask==3.0.0
notion-client==2.2.1
pyyaml==6.0.1
requests==2.31.0
ollama==0.1.6
python-dotenv==1.0.0
